changed relative position, moved transforms from tile to coordinates
