+++ src/aind_data_schema/core/rig.py
@@ -7,6 +7,7 @@ from pydantic import Field, ValidationInfo, field_validator
 from typing_extensions import Annotated
 
 from aind_data_schema.base import AindCoreModel
+from aind_data_schema.models.coordinates import Axis, Origin
 from aind_data_schema.models.devices import (
     LIGHT_SOURCES,
     Calibration,
@@ -48,7 +49,7 @@ class Rig(AindCoreModel):
 
     _DESCRIBED_BY_URL = AindCoreModel._DESCRIBED_BY_BASE_URL.default + "aind_data_schema/core/rig.py"
     describedBy: str = Field(_DESCRIBED_BY_URL, json_schema_extra={"const": _DESCRIBED_BY_URL})
-    schema_version: Literal["0.2.11"] = Field("0.2.11")
+    schema_version: Literal["0.2.12"] = Field("0.2.12")
     rig_id: str = Field(..., description="room_stim apparatus_version", title="Rig ID")
     modification_date: date = Field(..., title="Date of modification")
     mouse_platform: MOUSE_PLATFORMS
@@ -76,6 +77,8 @@ class Rig(AindCoreModel):
         title="CCF coordinate transform",
         description="Path to file that details the CCF-to-lab coordinate transform",
     )
+    origin_point: Optional[Origin] = Field(None, title="Origin point for rig position transforms")
+    rig_axes: Optional[List[Axis]] = Field(default=[], title="Rig axes", min_length=3, max_length=3)
     modalities: Set[Modality.ONE_OF] = Field(..., title="Modalities")
     notes: Optional[str] = Field(None, title="Notes")
