+++ src/aind_data_schema/imaging/tile.py
@@ -6,6 +6,12 @@ from typing import List, Literal, Optional, Union
 from pydantic import Field
 
 from aind_data_schema.base import AindModel
+from aind_data_schema.models.coordinates import (
+    Affine3dTransform,
+    Rotation3dTransform,
+    Scale3dTransform,
+    Translation3dTransform,
+)
 from aind_data_schema.models.units import AngleUnit, PowerUnit, SizeUnit
 
 
@@ -28,44 +34,6 @@ class Channel(AindModel):
     dilation: Optional[int] = Field(None, title="Dilation (pixels)")
     dilation_unit: SizeUnit = Field(SizeUnit.PX, title="Dilation unit")
     description: Optional[str] = Field(None, title="Description")
-
-
-class CoordinateTransform(AindModel):
-    """Generic base class for coordinate transform subtypes"""
-
-    type: str = Field(..., title="transformation type")
-
-
-class Scale3dTransform(CoordinateTransform):
-    """Values to be vector-multiplied with a 3D position, equivalent to the diagonals of a 3x3 transform matrix.
-    Represents voxel spacing if used as the first applied coordinate transform.
-    """
-
-    type: Literal["scale"] = "scale"
-    scale: List[Decimal] = Field(..., title="3D scale parameters", min_length=3, max_length=3)
-
-
-class Translation3dTransform(CoordinateTransform):
-    """Values to be vector-added to a 3D position. Often needed to specify a Tile's origin."""
-
-    type: Literal["translation"] = "translation"
-    translation: List[Decimal] = Field(..., title="3D translation parameters", min_length=3, max_length=3)
-
-
-class Rotation3dTransform(CoordinateTransform):
-    """Values to be vector-added to a 3D position. Often needed to specify a Tile's origin."""
-
-    type: Literal["rotation"] = "rotation"
-    rotation: List[Decimal] = Field(..., title="3D rotation matrix values (3x3) ", min_length=9, max_length=9)
-
-
-class Affine3dTransform(CoordinateTransform):
-    """Values to be vector-added to a 3D position. Often needed to specify a Tile's origin."""
-
-    type: Literal["affine"] = "affine"
-    affine_transform: List[Decimal] = Field(
-        ..., title="Affine transform matrix values (top 3x4 matrix)", min_length=12, max_length=12
-    )
 
 
 class Tile(AindModel):
